requirements.txt
pandas
pytest
streamlit
spotipy
google-genai
python-dotenv
